move some methods inside the pinia store
