+++ src/modules/data/board/index.ts
@@ -1,15 +1,17 @@
 import { useBoardPermissions } from "./BoardPermissions.composable";
 import { useBoardStore } from "./Board.store";
-import { useCardState } from "./CardState.store";
+import { useCardState } from "./CardState.composable";
 import { useSharedBoardPageInformation } from "./BoardPageInformation.composable";
 import { useBoardFocusHandler } from "./BoardFocusHandler.composable";
 import { useContentElementState } from "./ContentElementState.composable";
 import { useEditMode, useSharedEditMode } from "./EditMode.composable";
 import * as boardActions from "./boardActions/actions";
 import { useBoardSocketApi } from "./socket/socket";
+import { useCardStore } from "./Cards.store";
 
 export {
 	useBoardStore,
+	useCardStore,
 	useBoardSocketApi,
 	useCardState,
 	useContentElementState,
